css/css-animations/parsing/animation-name-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: getComputedValue().animationName</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-name">
<meta name="assert" content="animation-name computes to lists containing identifier, string or none.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("animation-name", 'NONE', 'none');

test_computed_value("animation-name", 'foo');
test_computed_value("animation-name", 'Both');
test_computed_value("animation-name", 'ease-in');
test_computed_value("animation-name", 'infinite');
test_computed_value("animation-name", 'paused');
test_computed_value("animation-name", 'first, second, third');

test_computed_value("animation-name", '"string"');
test_computed_value("animation-name", '"multi word string"');
test_computed_value("animation-name", '"initial"');
test_computed_value("animation-name", '"---\\22---"', '\"---\\\"---\"');

test_computed_value("animation-name", '"initial", "None", None, Last', '"initial", "None", none, Last');
</script>
</body>
</html>
